Pregunta si descarga la última versión de chromedriver

+++ tests/descargar_chromedriver.sh
@@ -2,10 +2,20 @@
 
 SCRIPT=$(readlink -f "$0")
 DIR=$(dirname "$SCRIPT")
-echo "Descargando la última versión de ChromeDriver..."
-ZIP=chromedriver_linux64.zip
-VERSION=$(curl -s https://chromedriver.storage.googleapis.com/LATEST_RELEASE)
-curl -s -o $DIR/$ZIP https://chromedriver.storage.googleapis.com/$VERSION/$ZIP
-unzip -o $DIR/$ZIP -d $DIR
-rm $DIR/$ZIP
-
+CHROME="chromedriver"
+SN="S"
+if [ -f "$DIR/$CHROME" ]
+then
+    echo -n "El archivo $CHROME ya existe. ¿Desea descargar la última versión? (s/N): "
+    read SN
+    [ "$SN" = "s" ] && SN="S"
+fi
+if [ "$SN" = "S" ]
+    then
+    echo "Descargando la última versión de ChromeDriver..."
+    ZIP=chromedriver_linux64.zip
+    VERSION=$(curl -s https://chromedriver.storage.googleapis.com/LATEST_RELEASE)
+    curl -s -o $DIR/$ZIP https://chromedriver.storage.googleapis.com/$VERSION/$ZIP
+    unzip -o $DIR/$ZIP -d $DIR
+    rm $DIR/$ZIP
+fi
